Part one of massive overhaul to mob system. Does not compile, because overhaul is only partially complete. If you need a working version, use revision 107. Current and future changes include:
Complete redo of damage system, allowing for things like easily adding new types of damage as well as generally cleaner code.
Removing targeting system for organs, because it's unnecessary and obnoxious bloat. Humans still take damage to specific organs, but it's purely cosmetic now. People will probably make a big stink about this, but honestly the organ targeting system did not add anything of value to the game.
Reorganizing mobs to be in a more sensible hierarchy, which will allow for better code reuse and much easier addition of new mobs. This will mean less bugs and (potentially) more types of creatures in the future, which will be good news for the medical researchers.

All of this is mostly for maintainability, not specifically for new features. However, some might be incidentally added along the way.

git-svn-id: http://svn.slurm.us/public/spacestation13/branches/kurper@108 4efa6ab2-2d15-dd11-83a9-0030489379b8