rev189 - "Ragzilla's patch to let AI read papers and be disconnected from cameras by wirecutters.
- Camera Users can read papers on floor/table up to 1 square away. Cannot read
anything in a player's hand (presumably the player rolls up the paper, or holds it in
some manner that prevents the camera user from seeing it.
- Attempts by the Camera User to view a paper >1 square away, will still pop a window
but it'll pop up the *d out version (like when you try to read as a monkey)
- If you want to pass a note to the AI/Camera User privately, you can *use* the paper
on the camera if you're in range, this will hold the note up to the camera and pop a
window on the AI. Interesting side effect of this that while unintentional, kind of
makes sense, is that if the AI cancels out of viewing the camera, but doesn't switch
to another one, you can still show it the note (RP justification is that the AI is
multitasking and semi paying attention to the camera while it's working on other things).
- Camera Users viewing a camera will now be disconnected when some mexican^h traitor
comes along and uses wire cutters on the camera, the AI gets a message about a loss
of comms, people using camera stations get a message about a burst of static and get
kicked out of the camera computer."

git-svn-id: http://svn.slurm.us/public/spacestation13/branches/kurper@320 4efa6ab2-2d15-dd11-83a9-0030489379b8