fingerprint changes - not based directly on name (which would have made it nonrandom!), fingerprints are 5 digits max, fingerprint scanners actually work.
theoretically it's possible for two people to have the same fingerprints, since there's only 65536 possible prints now instead of 16^32 or whatever. Theoretically there ought to be a collision about every 100 games of 30ish people, I think. At some point a way to guarantee uniqueness would be nice, but for now I'm not worrying about it.

git-svn-id: http://svn.slurm.us/public/spacestation13/branches/kurper@442 4efa6ab2-2d15-dd11-83a9-0030489379b8