ragzilla's wire bundle patch: "generalized the wire code so we can extend it to more items, made wire bundles able to accept anything under /obj/item/weapon that sends r_signal (currently timer/radio signaller/prox and infra sensors)
new /datum/assembly/wirebundle supports up to 24 wires (and yes we can add more if we need to) so we can extend this to atmos sensors and fire panels (I'm thinking instead of these devices sensing immediately near them, we can add remote fire/atmos sensor objects that have a hidden/virtual wire (read: it only exists at either end for wirecutting) from the sensor to the panel so you can cut sensors, or cut the whole panel's power (update alarms to look for dead panels maybe?). or if you're really nefarious and we add a semi-automated fire extinguishing system (fire alarm == auto trigger CO2 dump in area), hook your remote signaller/timer up to the remote dump controls and watch the fun as the area gets sealed off and gassed.

Oh and the best part of the patch is that it adds var/assmaster (assembly master) to obj/item/weapon :3"

git-svn-id: http://svn.slurm.us/public/spacestation13/branches/kurper@464 4efa6ab2-2d15-dd11-83a9-0030489379b8